+++ gdbgui/src/js/store.js
@@ -315,8 +315,9 @@ const initial_store_data = {
     gdb_autocomplete_options: [],
 
     gdb_console_entries: [],
-    
-    project_home: initial_data.project_home
+
+    show_filesystem: false,
+    middle_panes_split_obj: {},
 }
 
 const KEYS_TO_NOT_LOG_CHANGES_IN_CONSOLE = [